fix: reduce AG Grid re-renders (#305)
